membenarkan API

+++ app/src/main/java/com/example/pujasdelivery/data/MenuWithTenantName.kt
@@ -2,7 +2,7 @@ package com.example.pujasdelivery.data
 
 data class MenuWithTenantName(
     val id: Int,
-    val tenantId: Int,
+    val tenantId: Long,
     val name: String,
     val price: Int,
     val description: String,
